Move Indicator Update
Updated the move indicator feature so that the marker will appear at a
fixed location underneath the interactable object when player clicks on
it OR when player holds down LMB and drags the mouse pointer on top of
the interactable. (Dragging with LMB down onto an interactable, now
leads the player to interact with the object)

Set the offset value in the interactable object's inspector to set the
position where the marker will appear. Offsets have been set for the
indoor objects and only the studious man npc outside.