Ending Trigger
Set up assets and code for triggering the ending of the game, and
transitioning to the win screen. The trigger is still using placeholder
art, and the music manager will require some work to be able to account
for repeat playthroughs after the ending.

Note:
- may need to manually set under build settings the scene
"MainMenuWithoutManagers" as the final, 6th scene
- renamed a few music manager related variables for clarity.